+++ cli/col.c
@@ -7,7 +7,7 @@
 int tman_cli_col(int argc, char **argv, struct tman_context *ctx)
 {
     char *col;
-    struct tman_args args;
+    struct tman_arg args;
     int i, c, status, showhelp, showlist;
     struct tman_option opt;
     const char *errfmt = "cannot move to column '%s': %s";