+++ lib/tman.h
@@ -20,14 +20,14 @@ struct tmanstruct {
     char fstate[TMANPATHSIZE + 1];      /* file to store tman task state */
 };
 
-struct tman_args;
+struct tman_arg;
 struct tman_hook;
 struct tman_base;
 struct tman_option;
 struct tman_custom;
 struct tman_context;
 
-struct tman_args {
+struct tman_arg {
     char *id;
     char *brd;
     char *prj;
@@ -86,63 +86,63 @@ int tman_mkfs(void);
 const char *tman_strerror(void);
 
 /* Input irguments functions.  */
-int tman_get_args(struct tman_args *args);
-int tman_check_arg_id(struct tman_args *args);
-int tman_check_arg_prj(struct tman_args *args);
+int tman_get_args(struct tman_arg *args);
+int tman_check_arg_id(struct tman_arg *args);
+int tman_check_arg_prj(struct tman_arg *args);
 
 /* Task ID functions.  */
-int tman_id_add(struct tman_context *ctx, struct tman_args *args,
+int tman_id_add(struct tman_context *ctx, struct tman_arg *args,
                 struct tman_option *options);
-int tman_id_col(struct tman_context *ctx, struct tman_args *args, char *tag,
+int tman_id_col(struct tman_context *ctx, struct tman_arg *args, char *tag,
                 struct tman_option *options);
-int tman_id_del(struct tman_context *ctx, struct tman_args *args,
+int tman_id_del(struct tman_context *ctx, struct tman_arg *args,
                 struct tman_option *options);
-int tman_id_find_by_desc(struct tman_context *ctx, struct tman_args *args,
+int tman_id_find_by_desc(struct tman_context *ctx, struct tman_arg *args,
                          char *descpatt, struct tman_option *options);
-int tman_id_list(struct tman_context *ctx, struct tman_args *args,
+int tman_id_list(struct tman_context *ctx, struct tman_arg *args,
                  struct tman_option *options);
-int tman_id_link(struct tman_context *ctx, struct tman_args *args,
+int tman_id_link(struct tman_context *ctx, struct tman_arg *args,
                  struct tman_option *options);
 int tman_id_move(struct tman_context *ctx, char *srcprj, char *dstprj,
                  char *srcid, char *dstid);
-int tman_id_prev(struct tman_context *ctx, struct tman_args *args,
+int tman_id_prev(struct tman_context *ctx, struct tman_arg *args,
                  struct tman_option *options);
-int tman_id_set(struct tman_context *ctx, struct tman_args *args,
+int tman_id_set(struct tman_context *ctx, struct tman_arg *args,
                 struct unit *unitbin, struct tman_option *options);
-int tman_id_show(struct tman_context *ctx, struct tman_args *args,
+int tman_id_show(struct tman_context *ctx, struct tman_arg *args,
                  struct tman_option *options);
-int tman_id_sync(struct tman_context *ctx, struct tman_args *args,
+int tman_id_sync(struct tman_context *ctx, struct tman_arg *args,
                  struct tman_option *options);
-char *tman_id_getcurr(struct tman_context *ctx, struct tman_args *args);
-char *tman_id_getprev(struct tman_context *ctx, struct tman_args *args);
+char *tman_id_getcurr(struct tman_context *ctx, struct tman_arg *args);
+char *tman_id_getprev(struct tman_context *ctx, struct tman_arg *args);
 
 /* Task project functions.  */
-int tman_prj_add(struct tman_context *ctx, struct tman_args *args,
+int tman_prj_add(struct tman_context *ctx, struct tman_arg *args,
                  struct tman_option *options);
-int tman_prj_del(struct tman_context *ctx, struct tman_args *args,
+int tman_prj_del(struct tman_context *ctx, struct tman_arg *args,
                  struct tman_option *options);
 int tman_prj_list(struct tman_context *ctx, struct tman_option *options);
 int tman_prj_prev(struct tman_context *ctx, struct tman_option *options);
-int tman_prj_set(struct tman_context *ctx, struct tman_args *args,
+int tman_prj_set(struct tman_context *ctx, struct tman_arg *args,
                  struct tman_option *options);
-int tman_prj_sync(struct tman_context *ctx, struct tman_args *args,
+int tman_prj_sync(struct tman_context *ctx, struct tman_arg *args,
                   struct tman_option *options);
 char *tman_prj_getcurr(struct tman_context *ctx);
 char *tman_prj_getprev(struct tman_context *ctx);
 
 /* Task plugin functions.  */
 int tman_ispgn(char *pgndir, const char *pgname);
-int tman_pgnexec(struct tman_context *ctx, struct tman_args *args, char *pgname,
+int tman_pgnexec(struct tman_context *ctx, struct tman_arg *args, char *pgname,
                  char *pgncmd, struct tman_option *options);
 
 int tman_hook_action(struct tman_context *ctx, struct tman_hook *hooks,
-                     struct tman_args *args, char *cmd);
+                     struct tman_arg *args, char *cmd);
 struct unit *tman_hook_show(struct tman_context *ctx, struct tman_hook *hooks,
-                            struct tman_args *args, char *cmd);
+                            struct tman_arg *args, char *cmd);
 
-int tman_hook_action_free(struct tman_context *ctx, struct tman_args *args,
+int tman_hook_action_free(struct tman_context *ctx, struct tman_arg *args,
                           char *cmd);
 struct unit *tman_hook_show_free(struct tman_context *ctx,
-                                 struct tman_args *args);
+                                 struct tman_arg *args);
 
 #endif